Optimisations, round 1: Deal with tiles more efficiently
Instead of filling animation structures with both numbered images (i.e.
tiles) and named images, the latter of which are not animated, treat
numbered images differently.

Because only numbered images are now animatable, and their indices are
contiguous (from 0 to 255 inclusive), use list objects instead of dict
objects for efficiency.

Sprites are special, in that they may reference either type of image.
Arrange for this to work as before.