Optimisations, round 2: Draw the background with 1 blit call
Instead of drawing the background with 1 blit call per tile, compute
the background as a surface with alpha, then blit a viewport-sized
part of that surface with 1 blit call per frame.

I am optimising this code with the assumption that a single blit call
takes a thousand machine instructions to set up in Python, but the
blit is carried out by machine code in SDL which is much faster.