Optimisations, round 7: More immutable tiles and decoration changes
Tiles are now immutable if their appearance never changes, regardless
of whether they participate in collision detection or not. This brings
the number of immutable tiles drawn in the background layer to 124.

There is now a distinction between a decoration tile (one which is
immutable and doesn't participate in collision detection) and an
immutable tile. Decoration tiles get no entry in the table used for
collision detection, so as to speed up collision detection in mid-air.