Remove the high score from the Menu and prevent its creation
The high score presents a perverse incentive which rewards death
instead of good gameplay.

Vegetables are the only thing that gives score in this game. You can
collect them for points, and they respawn when you die.

Levels such as Volcano-2, which contain vegetables as well as lives
and coins (100 of which create a life), can create a situation where
the player extends the duration of his/her game indefinitely and can
get millions of points with ease by repeatedly gathering lives and
vegetables, dying then re-entering the level with all vegetables (and
coins, but those don't give points) intact.

(The life farming aspect is explicitly unhandled. It allows crafty
players who have difficulty with the Boss level to make their life
easier. Pun intended.)

This is in line with Mario platformers from the SNES era and later,
which keep score but not a high score.